requirements.txt
numpy==1.21.5
pandas==1.3.5
PyYAML==6.0
